+++ src/de.rs
@@ -40,19 +40,19 @@ use super::value::Value;
 /// `Number`, `String` and `Container`. They have three different decode methods.
 /// 1. `Null`, `True` and `False` can be obtained by `JEntry`, no extra work required.
 /// 2. `Number` and `String` has related `RawData`, `JEntry` store the length
-/// or offset of this data, the `Value` can be read out and then decoded.
+///    or offset of this data, the `Value` can be read out and then decoded.
 /// 3. `Container` is actually a nested `Array` or `Object` with the same structure,
-/// `JEntry` store the length or offset of the lower-level `Header`,
-/// from where the same decode process can begin.
-
-/// `RawData` is the encoded `Value`.
-/// `Number` is a variable-length `Decimal`, store both int and float value.
-/// `String` is the original string, can be borrowed directly without extra decode.
-/// `Array` and `Object` is a lower-level encoded `JSONB` value.
-/// The upper-level doesn't care about the specific content.
-/// Decode can be executed recursively.
-
-/// Decode `JSONB` Value from binary bytes.
+///    `JEntry` store the length or offset of the lower-level `Header`,
+///    from where the same decode process can begin.
+///
+///    `RawData` is the encoded `Value`.
+///    `Number` is a variable-length `Decimal`, store both int and float value.
+///    `String` is the original string, can be borrowed directly without extra decode.
+///    `Array` and `Object` is a lower-level encoded `JSONB` value.
+///    The upper-level doesn't care about the specific content.
+///    Decode can be executed recursively.
+///
+///    Decode `JSONB` Value from binary bytes.
 pub fn from_slice(buf: &[u8]) -> Result<Value<'_>, Error> {
     let mut decoder = Decoder::new(buf);
     match decoder.decode() {
@@ -122,7 +122,7 @@ impl<'a> Decoder<'a> {
             }
             NUMBER_TAG => {
                 let offset = jentry.length as usize;
-                let n = Number::decode(&self.buf[..offset]);
+                let n = Number::decode(&self.buf[..offset])?;
                 self.buf = &self.buf[offset..];
                 Ok(Value::Number(n))
             }