+++ src/lib.rs
@@ -73,6 +73,7 @@ mod iterator;
 mod jentry;
 pub mod jsonpath;
 pub mod keypath;
+mod lazy_value;
 mod number;
 mod parser;
 mod ser;
@@ -84,6 +85,8 @@ pub use error::Error;
 #[allow(unused_imports)]
 pub use from::*;
 pub use functions::*;
+pub use lazy_value::*;
 pub use number::Number;
+pub use parser::parse_lazy_value;
 pub use parser::parse_value;
 pub use value::*;